Struct labrpc::Network

pub struct Network {
    core: Arc<NetworkCore>,
}

Fields
core: Arc<NetworkCore>
Implementations

impl Network

pub fn new() -> Network

pub fn create() -> (Network, UnboundedReceiver<Rpc>)

fn start(&self, incoming: UnboundedReceiver<Rpc>)

pub fn add_server(&self, server: Server)

pub fn delete_server(&self, name: &str)

pub fn create_client(&self, name: String) -> Client

pub fn connect(&self, client_name: &str, server_name: &str)

Connects a Client to a server. a Client can only be connected once in its lifetime.

pub fn enable(&self, client_name: &str, enabled: bool)

Enable/disable a Client.

pub fn set_reliable(&self, yes: bool)

pub fn set_long_reordering(&self, yes: bool)

pub fn set_long_delays(&self, yes: bool)

pub fn count(&self, server_name: &str) -> usize

pub fn total_count(&self) -> usize

fn end_info(&self, client_name: &str) -> EndInfo

fn is_server_dead(
&self,
client_name: &str,
server_name: &str,
server_id: usize
) -> bool

async fn process_rpc(&self, rpc: Rpc) -> Result<Vec<u8>>

pub fn spawn<F>(&self, f: F) where
F: Future<Output = ()> + Send + 'static,

Spawns a future to run on this net framework.

pub fn spawn_poller<F>(&self, f: F) where
F: Future<Output = ()> + Send + 'static,

Spawns a future to run on this net framework.